refactor(rich-md-editor): freeze the formatting toolbar on scroll and extract the shared toolbar button
